+++ .eslintrc
@@ -13,3 +13,4 @@ extends:
 
 rules:
   semi: "error"
+  comma-dangle: 0
